경매 상품 등록 백엔드 코딩. 오류 수정 중.

+++ src/main/java/com/hagglemarket/marketweb/post/repository/PostImageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PostImageRepository extends JpaRepository<PostImage, Long> {
+public interface PostImageRepository extends JpaRepository<PostImage, Integer> {
     List<PostImage> findByPost(Post post);
 }
